pyproject.toml
[tool.poetry]
name = "flipdot"
version = "0.0.5"
description = ""
authors = ["Drew McDonald <github.com/drewmcdonald>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
freetype-py = "^2.5.1"
numpy = "^2.2.2"
pyserial = "^3.5"
fastapi = "^0.115.7"
pydantic = "^2.10.6"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pendulum = "^3.0.0"
aiofiles = "^23.2.1"
structlog = "^25.3.0"


[tool.poetry.group.dev.dependencies]
typing-extensions = "^4.12.2"
isort = "^5.13.2"
black = "^24.10.0"
types-requests = "^2.32.0.20241016"
mypy = "^1.15.0"
ruff = "^0.9.6"
types-pyserial = "^3.5.0.20250130"
watchfiles = "^1.0.4"
pytest = "^8.3.5"
numpy = "^2.2.6"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
flipdot-server = "flipdot.cli.server:run"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
lint.ignore = []
